# Conflicts:
#	demisto_sdk/commands/create_artifacts/content_creator.py
#	demisto_sdk/commands/create_artifacts/tests/content_creator_test.py

+++ demisto_sdk/commands/create_artifacts/content_creator.py
@@ -78,6 +78,7 @@ class ContentCreator:
             REPORTS_DIR,
             SCRIPTS_DIR,
             WIDGETS_DIR,
+            RELEASE_NOTES_DIR
         ]
 
         self.packages_to_skip = []
@@ -386,21 +387,24 @@ class ContentCreator:
 
             print(f"Finished process for {count_files} files")
 
-    def copy_dir_files(self, *args):
+    def copy_dir_files(self, *args, is_legacy_bundle=True):
         """
         Copy the yml, md, json and zip files from inside a directory to a bundle.
 
         :param args: (source directory, destination bundle)
+        :param is_legacy_bundle: flag to copy md files to bundle.
+            Should be False for content-new.zip and test-content.zip
         :return: None
         """
         # handle *.json files
         self.copy_dir_json(*args)
         # handle *.yml files
         self.copy_dir_yml(*args)
-        # handle *.md files
-        self.copy_dir_md(*args)
         # handle *.zip files
         self.add_tools_to_bundle(*args)
+        if not is_legacy_bundle:
+            # handle *.md files
+            self.copy_dir_md(*args)
 
     def copy_test_files(self, test_playbooks_dir=TEST_PLAYBOOKS_DIR):
         """
@@ -415,9 +419,10 @@ class ContentCreator:
             if os.path.isdir(path):
                 non_circle_tests = glob.glob(os.path.join(path, '*'))
                 for new_path in non_circle_tests:
-                    if self.check_from_version(new_path):
-                        print(f'Copying path {new_path}')
-                        new_file_path = self.add_suffix_to_file_path(os.path.join(self.test_bundle, os.path.basename(new_path)))
+                    if os.path.isfile(new_path) and self.check_from_version(new_path):
+                        print(f'copying path {new_path}')
+                        new_file_path = self.add_suffix_to_file_path(os.path.join(self.test_bundle,
+                                                                                  os.path.basename(new_path)))
                         shutil.copyfile(new_path, new_file_path)
                         self.add_from_version_to_yml(new_file_path)
 
@@ -521,7 +526,7 @@ class ContentCreator:
                             self.add_from_version_to_yml(new_file_path)
 
                 else:
-                    self.copy_dir_files(content_dir, dest_dir)
+                    self.copy_dir_files(content_dir, dest_dir, is_legacy_bundle=False)
 
     def update_content_version(self, content_ver: str = '', path: str = ''):
         regex = r'CONTENT_RELEASE_VERSION = .*'
@@ -644,7 +649,9 @@ class ContentCreator:
             shutil.make_archive(self.packs_zip, 'zip', self.packs_bundle)
 
             self.copy_file_to_artifacts('release-notes.md')
-            print_success(f'\nfinished creating the content artifacts at "{os.path.abspath(self.artifacts_path)}"')
+            self.copy_file_to_artifacts('beta-release-notes.md')
+            self.copy_file_to_artifacts('packs-release-notes.md')
+            print_success(f'finished creating the content artifacts at "{os.path.abspath(self.artifacts_path)}"')
         finally:
             if not self.preserve_bundles:
                 if os.path.exists(self.content_bundle):
